functions/gcs-file-sanitizer/requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
fsspec==0.7.4
gcsfs==0.6.2
google-api-core==1.22.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-crc32c==0.1.0
google-resumable-media==0.6.0
googleapis-common-protos==1.52.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
numpy==1.19.1
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.6
PyPDF2==1.26.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
xmltodict==0.12.0
